+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to Biology Dictionary will be documented in this file.
 
+## 1.0.2
+
+### Changed
+
+- Refactored Biology Dictionary screen scaling to improve compatibility with other mods.
+- Moved demo mode out of config and behind a debug-only hidden toggle on the About screen.
+- Moved debug mode to a static variable.
+
+### Fixed
+
+- Fixed beehive screen rendering.
+
 ## 1.0.1
 
 ### Added